Modify HIO system call prototypes to match the Linux syscall prototypes. This gets rid of some difficult to debug sign extention bugs on syscall return values.
All Linux system call prototypes return a long value.
hio_connect() was returning an int, which was being sign extended to
long and causing -EINPROGRESS failures to show up as a very large
negative number.  Glibc treats anything outside -4095 to 0 as
success and in for success does not set errno.  Cray libraries
were recognizing an error because return value was < 0 but
errno was set to 0, leading to confusion.

+++ kernel/hio/hio_syscalls/hio_close.c
@@ -6,11 +6,11 @@
 #include <lwk/hio.h>
 #include <lwk/aspace.h>
 
-extern int
-sys_close(int fd);
+extern long
+sys_close(unsigned int fd);
 
-int
-hio_close(int fd)
+long
+hio_close(unsigned int fd)
 {
 	if ( (!syscall_isset(__NR_close, current->aspace->hio_syscall_mask)) ||
 	     (fdTableFile(current->fdTable, fd))
